refac(message-reference): introduce own data schema builder
- moved message interpretation logic from builder to data_schema_builder together with the corresponding tests to data_schema_builder_test.go
- Call generateDataSchema from GenerateTDfromProtoBuf
